requirements.txt
fastapi==0.111.0
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.20.0
asyncio==3.4.3
bcrypt==4.1.3
pillow==10.4.0
pydantic==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
uvicorn[standard]==0.29.0